Show network error message
Instead of showing broken applications when there is no available
network, we now show an error message indicating that there is no
network available.  This is neat because now we dont have to let
customers know why things are working at shows, and people trying this
for themselves will know what they need to do to remedy the issue.

Change-Id: Ie447b129d64a559826196808a5b5489ae29cae5f
Reviewed-by: Lasse Räihä <lasse.raiha@digia.com>